Reproduction Attempt: Bug Fixes and Improvements Commit Series 6
Added sleep function to suspend the execution of accessing additional repository information using GitHub API for three seconds, due to 'API rate limit exceeded for user' error in r3_github_api.py.
Added check in the get_repo_info_github_api function to see if the information of the repository is already collected and stored in the database for a particular repository associated with an article.